Move the TextureShader class to the Velaptor.OpenGL.Shaders namespace and folder
