+++ Velaptor/OpenGL/Shaders/TextureShader.cs
@@ -2,7 +2,7 @@
 // Copyright (c) KinsonDigital. All rights reserved.
 // </copyright>
 
-namespace Velaptor.OpenGL
+namespace Velaptor.OpenGL.Shaders
 {
     // ReSharper disable RedundantNameQualifier
     using System;
